--- src/commands/init/index.ts
@@ -1,10 +1,0 @@
-import type { CommandModule } from 'yargs';
-import { init } from './init.js';
-
-export const command: CommandModule<{}> = {
-  command: ['init'],
-  describe: 'Automates basic configuration of AI CLI.',
-  handler: () => {
-    init();
-  },
-};
