--- src/interface/init/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { render } from 'ink';
-import { InitInterface } from './init-interface.js';
-
-type InitProps = {
-  configExists: boolean;
-};
-
-export const renderInitInterface = ({ configExists }: InitProps) => {
-  render(<InitInterface configExists={configExists} />);
-};
